refactor: ConsentsCommunication to split out state
